+++ neo/Implementations/Wallets/peewee/test_user_wallet.py
@@ -8,6 +8,9 @@ from neocore.UInt160 import UInt160
 from neocore.Fixed8 import Fixed8
 from neocore.KeyPair import KeyPair
 from neo.Wallets.NEP5Token import NEP5Token
+from neo.SmartContract.ContractParameterContext import ContractParametersContext
+from neo.Core.TX.Transaction import ContractTransaction, TransactionOutput
+from neo.Network.NodeLeader import NodeLeader
 import json
 import binascii
 
@@ -139,7 +142,7 @@ class UserWalletTestCase(WalletFixtureTestCase):
         contract = Blockchain.Default().GetContract(token_hash)
 
         token = NEP5Token(binascii.hexlify(contract.Code.Script))
-        token.Query(wallet)
+        token.Query()
 
         self.assertEqual(token.name, 'NEP5 Standard')
         self.assertEqual(token.decimals, 8)
@@ -172,3 +175,19 @@ class UserWalletTestCase(WalletFixtureTestCase):
         self.assertIsInstance(presult, bytearray)
 
         self.assertEqual(presult, self.wallet_1_script_hash.Data)
+
+    def test_9_send_neo_tx(self):
+
+        wallet = self.GetWallet1()
+
+        tx = ContractTransaction()
+        tx.outputs = [TransactionOutput(Blockchain.SystemShare().Hash, Fixed8.FromDecimal(10.0), self.import_watch_addr)]
+
+        tx = wallet.MakeTransaction(tx)
+
+        cpc = ContractParametersContext(tx)
+        wallet.Sign(cpc)
+        tx.scripts = cpc.GetScripts()
+
+        result = NodeLeader.Instance().Relay(tx)
+        self.assertEqual(result, True)