+++ neo/Settings.py
@@ -10,6 +10,8 @@ import json
 import os
 import logging
 from json.decoder import JSONDecodeError
+
+from neo import __version__
 from neocore.Cryptography import Helper
 
 import logzero
@@ -27,6 +29,7 @@ FILENAME_PREFERENCES = os.path.join(DIR_PROJECT_ROOT, 'preferences.json')
 FILENAME_SETTINGS_MAINNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.mainnet.json')
 FILENAME_SETTINGS_TESTNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.testnet.json')
 FILENAME_SETTINGS_PRIVNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.privnet.json')
+FILENAME_SETTINGS_COZNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.coz.json')
 
 
 class SettingsHolder:
@@ -47,18 +50,21 @@ class SettingsHolder:
 
     LEVELDB_PATH = None
     NOTIFICATION_DB_PATH = None
+
+    RPC_PORT = None
     NODE_PORT = None
     WS_PORT = None
     URI_PREFIX = None
-    VERSION_NAME = None
     BOOTSTRAP_FILE = None
+    NOTIF_BOOTSTRAP_FILE = None
 
     ALL_FEES = None
-
     USE_DEBUG_STORAGE = False
 
+    VERSION_NAME = "/NEO-PYTHON:%s/" % __version__
+
     # Logging settings
-    log_smart_contract_events = True
+    log_smart_contract_events = False
 
     # Helpers
     @property
@@ -72,6 +78,11 @@ class SettingsHolder:
         return self.NODE_PORT == 20333 and self.MAGIC == 1953787457
 
     @property
+    def is_coznet(self):
+        """ Returns True if settings point to CoZnet """
+        return self.NODE_PORT == 20333 and self.MAGIC == 1010102
+
+    @property
     def net_name(self):
         if self.MAGIC is None:
             return 'None'
@@ -79,6 +90,8 @@ class SettingsHolder:
             return 'MainNet'
         if self.is_testnet:
             return 'TestNet'
+        if self.is_coznet:
+            return 'CozNet'
         return 'PrivateNet'
 
     # Setup methods
@@ -103,11 +116,13 @@ class SettingsHolder:
 
         config = data['ApplicationConfiguration']
         self.LEVELDB_PATH = os.path.join(DIR_PROJECT_ROOT, config['DataDirectoryPath'])
+        self.RPC_PORT = int(config['RPCPort'])
         self.NODE_PORT = int(config['NodePort'])
         self.WS_PORT = config['WsPort']
         self.URI_PREFIX = config['UriPrefix']
-        self.VERSION_NAME = config['VersionName']
+
         self.BOOTSTRAP_FILE = config['BootstrapFile']
+        self.NOTIF_BOOTSTRAP_FILE = config['NotificationBootstrapFile']
 
         Helper.ADDRESS_VERSION = self.ADDRESS_VERSION
 
@@ -128,6 +143,10 @@ class SettingsHolder:
     def setup_privnet(self):
         """ Load settings from the privnet JSON config file """
         self.setup(FILENAME_SETTINGS_PRIVNET)
+
+    def setup_coznet(self):
+        """ Load settings from the coznet JSON config file """
+        self.setup(FILENAME_SETTINGS_COZNET)
 
     def set_log_smart_contract_events(self, is_enabled=True):
         self.log_smart_contract_events = is_enabled