+++ neo/SmartContract/ContractParameterContext.py
@@ -22,11 +22,16 @@ class ContractParamater():
     Value = None
 
     def __init__(self, type):
-        self.Type = type
+        if isinstance(type, ContractParameterType):
+            self.Type = type
+        elif isinstance(type, int):
+            self.Type = ContractParameterType(type)
+        else:
+            raise Exception("Invalid Contract Parameter Type %s. Must be ContractParameterType or int" % type)
 
     def ToJson(self):
         jsn = {}
-        jsn['type'] = ToName(self.Type)
+        jsn['type'] = self.Type.name
         return jsn
 
 
@@ -117,8 +122,6 @@ class ContractParametersContext():
         item = self.CreateItem(contract)
         item.ContractParameters[index].Value = parameter
 
-#        pdb.set_trace()
-
         return True
 
     def CreateItem(self, contract):
@@ -176,17 +179,16 @@ class ContractParametersContext():
             return True
 
         else:
-
             index = -1
+            if contract.ParameterList == '00':
+                contract.ParameterList = b'\x00'
             length = len(contract.ParameterList)
             for i in range(0, length):
-
-                if contract.ParameterList[i] == ContractParameterType.Signature:
+                if ContractParameterType(contract.ParameterList[i]) == ContractParameterType.Signature:
                     if index >= 0:
                         raise Exception("Signature must be first")
                     else:
                         index = i
-
             return self.Add(contract, index, signature)
 
     def GetIndex(self, script_hash):