+++ neo/SmartContract/ContractParameterType.py
@@ -5,9 +5,14 @@ Description:
 Usage:
     from neo.SmartContract.ContractParameterType import ContractParameterType
 """
+from enum import Enum
+import binascii
 
 
-class ContractParameterType(object):
+class ContractParameterType(Enum):
+    """
+    Contract Parameter Types are used to denote different types of objects used in the VM
+    """
     Signature = 0x00        # 签名
     Boolean = 0x01
     Integer = 0x02          # 整数
@@ -17,14 +22,38 @@ class ContractParameterType(object):
     PublicKey = 0x06
     String = 0x07
     Array = 0x10
+    InteropInterface = 0xf0
     Void = 0xff
+
+    def __str__(self):
+        return str(self.value.to_bytes(1, 'little').hex())
+
+    @staticmethod
+    def FromString(val):
+        """
+        Create a ContractParameterType object from a str
+
+        Args:
+            val (str): the value to be converted to a ContractParameterType
+
+        Returns:
+            ContractParameterType
+        """
+        return ContractParameterType(int.from_bytes(binascii.unhexlify(val), 'little'))
 
 
 import inspect
 
 
 def ToName(param_type):
+    """
+    Gets the name of a ContractParameterType based on its value
+    Args:
+        param_type (ContractParameterType): type to get the name of
 
+    Returns:
+        str
+    """
     items = inspect.getmembers(ContractParameterType)
 
     if type(param_type) is bytes: