+++ neo/SmartContract/StateReader.py
@@ -176,7 +176,13 @@ class StateReader(InteropService):
     def ExecutionCompleted(self, engine, success, error=None):
 
         height = Blockchain.Default().Height
-        tx_hash = engine.ScriptContainer.Hash
+        tx_hash = None
+
+        if engine.ScriptContainer:
+            tx_hash = engine.ScriptContainer.Hash
+
+        if not tx_hash:
+            tx_hash = UInt256(data=bytearray(32))
 
         entry_script = None
         try:
@@ -229,6 +235,8 @@ class StateReader(InteropService):
         return True
 
     def CheckWitnessHash(self, engine, hash):
+        if not engine.ScriptContainer:
+            return False
 
         if self._hashes_for_verifying is None:
             container = engine.ScriptContainer
@@ -285,12 +293,17 @@ class StateReader(InteropService):
 
         hash = UInt160(data=engine.CurrentContext.ScriptHash())
 
+        tx_hash = None
+
+        if engine.ScriptContainer:
+            tx_hash = engine.ScriptContainer.Hash
+
         # Build and emit smart contract event
         self.events_to_dispatch.append(SmartContractEvent(SmartContractEvent.RUNTIME_LOG,
                                                           [message],
                                                           hash,
                                                           Blockchain.Default().Height,
-                                                          engine.ScriptContainer.Hash,
+                                                          tx_hash,
                                                           test_mode=engine.testMode))
 
         return True