+++ neo/VM/ExecutionEngine.py
@@ -93,7 +93,7 @@ class ExecutionEngine():
 
     def ResultsForCode(self, contract):
         try:
-            return_type = contract.ReturnType
+            return_type = ContractParameterType(contract.ReturnType)
 
             item = self.EvaluationStack.Items[0]
             if return_type == ContractParameterType.Integer:
@@ -598,8 +598,7 @@ class ExecutionEngine():
 
                 except Exception as e:
                     estack.PushT(False)
-                    traceback.print_stack()
-                    traceback.print_exc()
+                    logger.error("Could not checksig: %s " % e)
 
             elif opcode == CHECKMULTISIG:
 
@@ -624,7 +623,7 @@ class ExecutionEngine():
                 for i in range(0, m):
                     sigs.append(estack.Pop().GetByteArray())
 
-                message = self.ScriptContainer.GetMessage()
+                message = self.ScriptContainer.GetMessage() if self.ScriptContainer else ''
 
                 fSuccess = True