+++ CHANGELOG.rst
@@ -1,12 +1,26 @@
 Changelog
 =========
 
-All notable changes to this project following the ``0.4.0`` release will be documented in this file.
+All notable changes to this project are documented in this file.
 
-[0.4.7-dev] work in progress
+[0.4.9-dev] work in progress
 ----------------------------
 
-- ...
+- Close wallet on quit (`PR #226 <https://github.com/CityOfZion/neo-python/pull/226>`_)
+- Bugfix for smart contract storage events (`PR #228 <https://github.com/CityOfZion/neo-python/pull/228>`_)
+
+
+[0.4.8] 2018-02-15
+------------------
+
+- Fix Gas Cost Calculation (`PR #220 <https://github.com/CityOfZion/neo-python/pull/220>`_)
+- Clarify message for token mint command (`PR #212 <https://github.com/CityOfZion/neo-python/pull/212>`_)
+- Troubleshooting osx script (`PR #208 <https://github.com/CityOfZion/neo-python/pull/208>`_)
+- Make Contract Search case insensitive (`PR #207 <https://github.com/CityOfZion/neo-python/pull/207>`_)
+- implement a more robust CLI command parser
+- added peristence to NotificationDB for NEP5 Tokens
+- upstream neocore update
+
 
 [0.4.6] 2018-01-24
 ------------------
@@ -16,6 +30,7 @@ All notable changes to this project following the ``0.4.0`` release will be docu
 - Added ability to include transaction attributes with the send command. example: ``send neo APRgMZHZubii29UXF9uFa6sohrsYupNAvx 10 --tx-attr={'usage':241,'data':'My Remark'}`` (`PR #184 <https://github.com/CityOfZion/neo-python/pull/184>`_)
 - Notification REST API (`PR #177 <https://github.com/CityOfZion/neo-python/pull/177>`_, `examples/notification-rest-api-server.py <https://github.com/CityOfZion/neo-python/blob/development/examples/notification-rest-api-server.py>`_)
 - Minor cleanups and documentation updates
+
 
 [0.4.5] 2018-01-18
 ------------------
@@ -29,11 +44,13 @@ All notable changes to this project following the ``0.4.0`` release will be docu
 - lots of documentation
 - various small bugfixes
 
+
 [0.4.3] 2017-12-21
 ------------------
 
 - updated ``neo-boa`` to ``0.2.1``
 - added support for array ``REVERSE`` and ``APPEND`` VM opcodes
+
 
 [0.4.3] 2017-12-21
 ------------------